cleanup(maven): consolidate BuildStateManager into shared module
Move BuildStateManager to shared module. NxMaven and NxMaven3 now look
up MavenProjectHelper themselves and pass it to the shared manager.
