+++ packages/maven/batch-runner/src/main/resources/nx-maven-adapters/maven3/classes.txt
@@ -1,4 +1,3 @@
-dev/nx/maven/adapter/maven3/BuildStateManager3.class
 dev/nx/maven/adapter/maven3/CachingMaven3Invoker.class
 dev/nx/maven/adapter/maven3/NxMaven3$getProjectMapMethod$2.class
 dev/nx/maven/adapter/maven3/NxMaven3.class
